+++ NEWS.md
@@ -1,6 +1,7 @@
 # queryparser (development version)
 
 * `BETWEEN` expressions with quotes in operands translate correctly ([#13](https://github.com/ianmcook/queryparser/issues/13))
+* Line comments (`--`) and block comments (`/* */`) are removed from queries ([#14](https://github.com/ianmcook/queryparser/issues/14))
 * Continuous integration and coverage tests
 * Minor bugfixes and improvements